requirements.txt
# python -m spacy download en_core_web_md

faiss-cpu==1.7.4
numpy==1.26.3
openai==1.8.0
pandas==2.1.4
PyMuPDF==1.23.15
PyAudio==0.2.14
scikit-learn==1.3.2
scipy==1.11.4
sentence-splitter==1.4
spacy==3.7.2
SpeechRecognition==3.10.1
SpeechRecognition
tiktoken==0.5.2
